+++ backend/canisters/community/impl/src/queries/selected_updates.rs
@@ -84,6 +84,12 @@ fn selected_updates_impl(args: Args, state: &RuntimeState) -> Response {
                     user_updates_handler.mark_member_updated(&mut result, *user_id, false, false);
                 }
             }
+            CommunityEventInternal::MemberLapsed(e) => {
+                user_updates_handler.mark_member_updated(&mut result, e.user_id, false, false);
+            }
+            CommunityEventInternal::MemberUnlapsed(e) => {
+                user_updates_handler.mark_member_updated(&mut result, e.user_id, false, false);
+            }
             CommunityEventInternal::UsersBlocked(e) => {
                 for user_id in e.user_ids.iter() {
                     let referral = is_my_referral(e.referred_by.get(user_id).copied(), &my_user_id);