+++ backend/canisters/community/impl/src/updates/c2c_join_community.rs
@@ -155,12 +155,7 @@ pub(crate) fn join_community_impl(args: &Args, state: &mut RuntimeState) -> Resu
             now,
         );
     } else {
-        state
-            .data
-            .members
-            .get_by_user_id_mut(&args.user_id)
-            .unwrap()
-            .clear_lapsed(now);
+        state.data.update_lapsed(args.user_id, None, false, now);
     }
 
     // If there is a payment gate on this community then queue payments to owner(s) and treasury